episode 22

+++ database/migrations/2020_06_05_081316_create_activities_table.php
@@ -15,7 +15,7 @@ class CreateActivitiesTable extends Migration
     {
         Schema::create('activities', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedInteger('project_id');
+            $table->unsignedBigInteger('project_id');
             $table->string('description');
             $table->timestamps();
